Made compatible with Swift Package Manager
mth + mthcolor targets exported

color.c was not used in the mth Makefile but still seemed useful, so
split into a separate target which can be imported into SPM projects.
